Refactor message and room serializer into own file serializer.py
